+++ src/FreeElephants/TSxUnit/DocumentTestCase.ts
@@ -1,5 +1,6 @@
 ///<reference path="Mock/DocumentBuilder.ts"/>
 ///<reference path="Assert/DomAssert.ts"/>
+///<reference path="Mock/DocumentEventBuilder.ts"/>
 /**
  * Base class for DOM interaction testing.
  */
@@ -7,6 +8,7 @@ namespace FreeElephants.TSxUnit {
 
     import DocumentBuilder = FreeElephants.TSxUnit.Mock.DocumentBuilder;
     import DomAssert = FreeElephants.TSxUnit.Assert.DomAssert;
+    import DocumentEventBuilder = FreeElephants.TSxUnit.Mock.DocumentEventBuilder;
 
     export abstract class DocumentTestCase extends TestCase {
 
@@ -21,9 +23,23 @@ namespace FreeElephants.TSxUnit {
             DomAssert.assertElementNotExists(doc, selector, this._undefined);
         }
 
+        protected assertElementHasClass(element: HTMLElement, className: string) {
+            this.numberOfAsserts++;
+            DomAssert.assertElementHasClass(element, className, this._undefined);
+        }
+
+        protected assertElementNotHasClass(element: HTMLElement, className: string) {
+            this.numberOfAsserts++;
+            DomAssert.assertElementNotHasClass(element, className, this._undefined);
+        }
+
         protected getFakeDocumentBuilder(): DocumentBuilder {
             return new DocumentBuilder();
         }
 
+
+        protected getFakeDocumentEventBuilder(document: Document, eventInterface: string, eventType: string) {
+            return new DocumentEventBuilder(document, eventInterface, eventType);
+        }
     }
 }